SF-798 Remove user docs and project associations (#896)
* SF-798 First add tests and prepare for delete-user implementation

- Add comments.
- Introduce IRealtimeServer to allow mocking.
- Add early-crash argument checking to some methods. This should be
the commit's only behavioural change.
- Expose RemoveUserFromProjectAsync() as internal for unit testing.
- Add tests for existing functionality.
- ProjectServiceTests Sites are expanded to be consistent with
TestProject documents, for these or follow-up tests.

* SF-798 Remove user docs and project associations

- When a sysadmin deletes a user, SF runs UserService
  DeleteAsync(), which calls:
  - ProjectService.RemoveUserFromAllProjectsAsync()
    which calls RemoveUserCoreAsync() for each user doc Sites.Projects
      which calls RemoveUserFromProjectAsync().
  - UserSecrets.DeleteAsync()
  - UserDoc.DeleteAsync()
  - RealtimeService.DeleteUserAsync(), removing docs from collections
    users, sf_project_user_configs, o_users,
    and o_sf_project_user_configs.
- RealtimeService DeleteUserAsync() removes docs, as matched by
  finding the userId at the end of the '_id' or 'd' field.
- The knowledge for what collections to remove user-related docs from
  (in addition to 'users') can be defined per-application, and here is
  done in SFRealtimeServiceCollectionExtensions.cs.

+++ src/SIL.XForge/Services/UserService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
@@ -22,14 +23,16 @@ namespace SIL.XForge.Services
         private readonly IOptions<SiteOptions> _siteOptions;
         private readonly IRepository<UserSecret> _userSecrets;
         private readonly IAuthService _authService;
+        private readonly IProjectService _projectService;
 
         public UserService(IRealtimeService realtimeService, IOptions<SiteOptions> siteOptions,
-            IRepository<UserSecret> userSecrets, IAuthService authService)
+            IRepository<UserSecret> userSecrets, IAuthService authService, IProjectService projectService)
         {
             _realtimeService = realtimeService;
             _siteOptions = siteOptions;
             _userSecrets = userSecrets;
             _authService = authService;
+            _projectService = projectService;
         }
 
         public async Task UpdateUserFromProfileAsync(string curUserId, string userProfileJson)
@@ -127,16 +130,29 @@ namespace SIL.XForge.Services
             }
         }
 
+        /// <summary>
+        /// Delete user with SF user id userId, as requested by SF user curUserId who has systemRole.
+        /// </summary>
         public async Task DeleteAsync(string curUserId, string systemRole, string userId)
         {
+            if (curUserId == null || systemRole == null || userId == null)
+            {
+                throw new ArgumentNullException();
+            }
             if (systemRole != SystemRole.SystemAdmin && userId != curUserId)
+            {
                 throw new ForbiddenException();
+            }
 
+            await _projectService.RemoveUserFromAllProjectsAsync(curUserId, userId);
+            await _userSecrets.DeleteAsync(userId);
             using (IConnection conn = await _realtimeService.ConnectAsync(curUserId))
             {
                 IDocument<User> userDoc = await conn.FetchAsync<User>(userId);
                 await userDoc.DeleteAsync();
             }
+            // Remove the actual docs.
+            await _realtimeService.DeleteUserAsync(userId);
         }
 
         /// <summary>
